refactor: rename KrispCall to krisp-call
