--- packages/pieces/community/KrispCall/README.md
@@ -1,7 +1,0 @@
-# pieces-KrispCall
-
-This library was generated with [Nx](https://nx.dev).
-
-## Building
-
-Run `nx build pieces-KrispCall` to build the library.
