--- packages/pieces/community/KrispCall/package.json
@@ -1,4 +1,0 @@
-{
-  "name": "@activepieces/piece-KrispCall",
-  "version": "0.0.1"
-}
