--- packages/pieces/community/KrispCall/project.json
@@ -1,38 +1,0 @@
-{
-  "name": "pieces-KrispCall",
-  "$schema": "../../../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "packages/pieces/community/KrispCall/src",
-  "projectType": "library",
-  "targets": {
-    "build": {
-      "executor": "@nx/js:tsc",
-      "outputs": [
-        "{options.outputPath}"
-      ],
-      "options": {
-        "outputPath": "dist/packages/pieces/community/KrispCall",
-        "tsConfig": "packages/pieces/community/KrispCall/tsconfig.lib.json",
-        "packageJson": "packages/pieces/community/KrispCall/package.json",
-        "main": "packages/pieces/community/KrispCall/src/index.ts",
-        "assets": [
-          "packages/pieces/community/KrispCall/*.md"
-        ],
-        "buildableProjectDepsInPackageJsonType": "dependencies",
-        "updateBuildableProjectDepsInPackageJson": true
-      }
-    },
-    "publish": {
-      "command": "node tools/scripts/publish.mjs pieces-KrispCall {args.ver} {args.tag}",
-      "dependsOn": [
-        "build"
-      ]
-    },
-    "lint": {
-      "executor": "@nx/eslint:lint",
-      "outputs": [
-        "{options.outputFile}"
-      ]
-    }
-  },
-  "tags": []
-}
