+++ packages/pieces/community/krisp-call/README.md
@@ -1,0 +1,7 @@
+# pieces-krisp-call
+
+This library was generated with [Nx](https://nx.dev).
+
+## Building
+
+Run `nx build pieces-krisp-call` to build the library.
