+++ packages/pieces/community/krisp-call/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "@activepieces/piece-krisp-call",
+  "version": "0.0.1"
+}
