+++ packages/pieces/community/krisp-call/src/index.ts
@@ -1,13 +1,12 @@
-
-import { createPiece, PieceAuth } from "@activepieces/pieces-framework";
+import { createPiece, PieceAuth } from '@activepieces/pieces-framework';
 import { httpClient, HttpMethod } from '@activepieces/pieces-common';
 import { addContact } from './lib/actions/add-contact';
-import { deleteContacts } from "./lib/actions/delete-contacts";
-import { sendSms } from "./lib/actions/send-sms";
-import { sendMms } from "./lib/actions/send-mms";
-import { newContact } from "./lib/triggers/new-contact";
-import { newMms } from "./lib/triggers/new-mms";
-import { newVoicemail } from "./lib/triggers/new-voicemail";
+import { deleteContacts } from './lib/actions/delete-contacts';
+import { sendSms } from './lib/actions/send-sms';
+import { sendMms } from './lib/actions/send-mms';
+import { newContact } from './lib/triggers/new-contact';
+import { newMms } from './lib/triggers/new-mms';
+import { newVoicemail } from './lib/triggers/new-voicemail';
 
 export const krispcallAuth = PieceAuth.CustomAuth({
   props: {
@@ -40,10 +39,11 @@ export type krispcallAuth = {
 };
 
 export const KrispCall = createPiece({
-  displayName: "Krispcall",
+  displayName: 'Krispcall',
   auth: krispcallAuth,
   minimumSupportedRelease: '0.20.0',
-  logoUrl: "https://krispcall.com/wp-content/uploads/2023/06/krispcall-favicon.svg",
+  logoUrl:
+    'https://krispcall.com/wp-content/uploads/2023/06/krispcall-favicon.svg',
   authors: [],
   actions: [addContact, deleteContacts, sendSms, sendMms],
   triggers: [newContact, newMms, newVoicemail],
