+++ packages/pieces/community/krisp-call/src/lib/actions/delete-contacts.ts
@@ -23,13 +23,8 @@ export const deleteContacts = createAction({
       },
       body: {
         contacts: propsValue.contacts,
-
-      }
+      },
     });
     return res.body;
   },
 });
-
-
-
-
