+++ packages/pieces/community/krisp-call/src/lib/actions/send-sms.ts
@@ -1,6 +1,4 @@
-import {
-  createAction,
-} from '@activepieces/pieces-framework';
+import { createAction } from '@activepieces/pieces-framework';
 import { httpClient, HttpMethod } from '@activepieces/pieces-common';
 import { krispcallAuth } from '../..';
 import { Property, PiecePropValueSchema } from '@activepieces/pieces-framework';
@@ -33,7 +31,7 @@ export const sendSms = createAction({
               'X-API-KEY': authVaue.apiKey,
             },
           });
-          const mappedOptions = res?.body?.map(item => {
+          const mappedOptions = res?.body?.map((item) => {
             return {
               label: item.name,
               value: item.number,
@@ -42,29 +40,28 @@ export const sendSms = createAction({
 
           return {
             disabled: false,
-            options: mappedOptions
+            options: mappedOptions,
           };
         } catch (error) {
           // Handle error
           console.error(error);
           return { disabled: true, options: [] }; // Return empty options array or handle error accordingly
         }
-      }
-    })
-    ,
+      },
+    }),
     to_number: Property.Number({
       displayName: 'To Number',
       description: 'Enter the number to which you want to send sms.',
-      required: true
+      required: true,
     }),
     content: Property.ShortText({
       displayName: 'content',
       description: 'Enter your message here.',
-      required: true
+      required: true,
     }),
   },
   async run({ auth, propsValue }) {
-    console.log(auth.apiKey)
+    console.log(auth.apiKey);
     const res = await httpClient.sendRequest<string[]>({
       method: HttpMethod.POST,
       url: 'https://automationapi.krispcall.com/api/v1/platform/activepiece/send-sms',
@@ -75,10 +72,9 @@ export const sendSms = createAction({
       body: {
         from_number: propsValue.from_number,
         to_number: propsValue.to_number,
-        content: propsValue.content
-      }
+        content: propsValue.content,
+      },
     });
     return res.body;
   },
 });
-
